Separate hardware display drawing engine from software
